Fix: Unchecked validation in UserUseCase class

+++ backend/src/main/java/com/unibague/magno/domain/usecase/UserUseCase.java
@@ -38,6 +38,10 @@ public class UserUseCase implements IUserServicePort {
 
     @Override
     public void deleteById(Long id) {
+        if (userPersistencePort.findById(id).isEmpty()) {
+            throw new UserNotFoundException(
+                    String.format("User with id %s could not be deleted because it does not exist", id));
+        }
         userPersistencePort.deleteById(id);
     }
